moving pseudop code into its own package
